+++ src/packing.c
@@ -16,11 +16,11 @@ unsigned long pack()
 {
     // Carrega arquivo a compactar no buffer (e seu tamanho em buffer_length)
     byte * buffer;
-    unsigned long buffer_length;
+    unsigned long long buffer_length;
     buffer = load_file_into_buffer(FILE_NAME_STR, &buffer_length);
 
     // Tamanho inicial do arquivo
-    printf("Lidos %lu bytes\n", buffer_length);
+    printf("Lidos %llu bytes\n", buffer_length);
 
     // Prepara nome de saida do arquivo compactado
     byte output[strlen(FILE_NAME_STR) + 6];
@@ -32,7 +32,8 @@ unsigned long pack()
     // Constroi arvore de huffman
     HufNode* tree_root = build_huffman_tree(list_head);
 
-    // Prepara pra montar a string representando a arvore em pre-ordem
+    /* Prepara pra montar uma array de bytes representando a arvore em pre-ordem.
+     * Essa array vai pro cabeçalho do arquivo compactado. */
     unsigned long tree_byte_arr_length = 4096;// Espaço mais que sufuciente para armazenar os bytes da arvore em string
     unsigned long tree_byte_arr_load = 0;
     byte tree_byte_arr[tree_byte_arr_length];
@@ -58,8 +59,10 @@ unsigned long pack()
 
     // Novo buffer pra receber material compactado
     byte* compressed_buffer;
+
     // Armazena tamanho do buffer depois de compactado
     unsigned long compressed_buffer_length;
+
     // Armazena quantos bits de lixo ficou no ultimo byte
     unsigned int last_byte_garbage;
 
@@ -88,7 +91,7 @@ byte* build_header(const byte* tree_byte_arr, unsigned long tree_byte_arr_length
 
     unsigned int garbage = last_byte_garbage << 5; // Põe os três bits na extremidade esquerda do byte
     unsigned int part_a = tree_byte_arr_length >> 8; // Elimina primeiro byte e põe o próximo na extremidade direita
-    unsigned int part_b = tree_byte_arr_length & (UINT_MAX - 65280); // Pega os bits definidos somente no primeiro byte de tree_byte_arr_length
+    unsigned int part_b = tree_byte_arr_length & (UINT_MAX - 65280U); // Pega os bits definidos somente no primeiro byte de tree_byte_arr_length
 
     // Os bytes além do primeiro são descartados nessa operação
     header[0] = part_a | garbage;
@@ -98,6 +101,7 @@ byte* build_header(const byte* tree_byte_arr, unsigned long tree_byte_arr_length
     for (unsigned long i=0; i < tree_byte_arr_length; i++)
         header[i+2] = tree_byte_arr[i];
 
+    // Debug
     dump_to_file("logs/header_str.log", header, 2, tree_byte_arr_length+2);
 
     return header;
@@ -153,10 +157,12 @@ byte* compress_byte_stream(const byte* stream, unsigned long stream_length,
         for (int i=0; i < strlen(code); i++) {
             state = 0;
 
+            // Define bit desejado como 0
             if (code[i] == '0') {
                 state &= ~(_1 << (7 - bit));
                 buffer[one_byte] &= ~(_1 << (7 - bit)); }
 
+            // Define bit desejado como 1
             else if (code[i] == '1') {
                 state |= (_1 << (7 - bit));
                 buffer[one_byte] |= (_1 << (7 - bit)); }
@@ -164,7 +170,7 @@ byte* compress_byte_stream(const byte* stream, unsigned long stream_length,
             // Atualiza qual byte de buffer está sendo escrito atualmente
             if (++bit > 7) { // Passou do último bit do byte atual
                 one_byte++; // Avança para o próximo byte
-                bit = 0; // Primeiro bit do próximo byte será definido no próximo ciclo
+                bit = 0; // Será definido o primeiro bit do próximo byte no próximo ciclo
                 *compressed_size += 1; } // Mais um byte será usado, aumentando o tamanho final em 1
         }
     }
@@ -172,7 +178,7 @@ byte* compress_byte_stream(const byte* stream, unsigned long stream_length,
 
     // Se bit == 0, temos o ultimo one_byte vazio e nenhum bit de lixo
     if (bit == 0) {
-        *compressed_size -= 1;
+        *compressed_size -= 1; // Isso porque a gente incrementa compressed_size cada vez que enche um byte
         *garbage_length = 0; }
     else
         *garbage_length = 8 - bit; // Calcula bits de lixo no ultimo byte
